// src/lib/types.ts
export interface SanityDocument {
  _id: string
  _type: string
  _createdAt?: string
  _updatedAt?: string
  _rev?: string
}

export interface Slug {
  _type: 'slug'
  current: string
}

export interface SanityImage {
  _type?: 'image'
  src?: string
  asset?: {
    _ref: string
    _type: 'reference'
  }
  alt?: string
}

export type DivisionType = 'civil' | 'electrical' | 'mechanical' | 'it'
// ...
export interface ProjectDivision {
  name: string
  type?: DivisionType
}

export interface ProjectListItem extends SanityDocument {
  _type: 'project'
  title: string
  titleAr?: string
  slug?: Slug
  shortDescription?: string
  shortDescriptionAr?: string
  thumbnail?: SanityImage
  division?: ProjectDivision
  tags?: string[]
  seoTitle?: string
  seoTitleAr?: string
  seoTitleAlt?: string
  seoTitleAltAr?: string
  seoDescription?: string
  seoDescriptionAr?: string
  seoDescriptionAlt?: string
  seoDescriptionAltAr?: string
  seoKeywords?: string
  seoKeywordsAr?: string
}